Move settings button component into its own directory
