+++ src/components/SettingsButton/SettingsButton.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import Button from '../../UI/atoms/Button/Button';
-import './Header.css';
 
 interface SettingsButtonProps {
   showSettings(): void;
